Move to patches module
